refactor: move base classes to related modules
move report and indicator base classes to indicators and reports modules
