+++ workers/ohsome_quality_analyst/indicators/mapping_saturation/indicator.py
@@ -9,8 +9,8 @@ from dateutil.parser import isoparse
 from geojson import Feature
 from rpy2.rinterface_lib.embedded import RRuntimeError
 
-from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.base.topic import BaseTopic as Topic
+from ohsome_quality_analyst.indicators.base import BaseIndicator
 from ohsome_quality_analyst.indicators.mapping_saturation import models
 from ohsome_quality_analyst.ohsome import client as ohsome_client
