+++ workers/ohsome_quality_analyst/indicators/minimal/indicator.py
@@ -4,8 +4,8 @@ from string import Template
 import dateutil.parser
 from geojson import Feature
 
-from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.base.topic import BaseTopic as Topic
+from ohsome_quality_analyst.indicators.base import BaseIndicator
 from ohsome_quality_analyst.ohsome import client as ohsome_client
