+++ workers/ohsome_quality_analyst/indicators/poi_density/indicator.py
@@ -7,9 +7,9 @@ import matplotlib.pyplot as plt
 import numpy as np
 from geojson import Feature
 
-from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.base.topic import BaseTopic as Topic
 from ohsome_quality_analyst.geodatabase.client import get_area_of_bpolys
+from ohsome_quality_analyst.indicators.base import BaseIndicator
 from ohsome_quality_analyst.ohsome import client as ohsome_client
 
 # threshold values defining the color of the traffic light
