+++ workers/ohsome_quality_analyst/reports/base.py
@@ -8,12 +8,12 @@ from typing import List, Literal, NamedTuple, Tuple
 import numpy as np
 from geojson import Feature
 
-from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.definitions import get_attribution, get_metadata
 from ohsome_quality_analyst.html_templates.template import (
     get_template,
     get_traffic_light,
 )
+from ohsome_quality_analyst.indicators.base import BaseIndicator
 from ohsome_quality_analyst.utils.helper import flatten_dict
