+++ workers/ohsome_quality_analyst/reports/multilevel_mapping_saturation/report.py
@@ -1,7 +1,7 @@
 from geojson import Feature
 
-from ohsome_quality_analyst.base.report import BaseReport, IndicatorTopic
 from ohsome_quality_analyst.definitions import get_attribution
+from ohsome_quality_analyst.reports.base import BaseReport, IndicatorTopic
 
 
 class MultilevelMappingSaturation(BaseReport):
